+++ web/ASC.Web.Client/src/components/pages/Wizard/sub-components/header-container.js
@@ -2,7 +2,12 @@ import React from 'react';
 import PropTypes from "prop-types";
 import styled from 'styled-components';
 
-import { Box, Heading, Text, utils } from 'asc-web-components';
+import { 
+  Box, 
+  Heading, 
+  Text, 
+  utils 
+} from 'asc-web-components';
 
 const { tablet } = utils.device;
